+++ example/firestore_test/bin/server.dart
@@ -17,7 +17,7 @@
 import 'package:firebase_functions/firebase_functions.dart';
 
 void main(List<String> args) async {
-  await fireUp(args, (firebase) {
+  await runFunctions((firebase) {
     // Test Firestore onDocumentCreated with wildcard
     firebase.firestore.onDocumentCreated(document: 'users/{userId}', (
       event,